+++ opensrp-vaksinator/src/main/java/util/formula/Support.java
@@ -156,6 +156,6 @@ public class Support {
         int tahun = Integer.parseInt(currentDate.substring(0,4))-Integer.parseInt(lastVisitDate.substring(0,4));
         int bulan = Integer.parseInt(currentDate.substring(5,7))-Integer.parseInt(lastVisitDate.substring(5,7));
         int hari = Integer.parseInt(currentDate.substring(8))-Integer.parseInt(lastVisitDate.substring(8));
-        return(tahun*12 + bulan + (int)(hari/30));
+        return(tahun*12 + bulan + hari/30);
     }
 }
